requirements.txt
openai
chromadb
langchain
tiktoken
python-dotenv
